feat(toast): consolidate withToast HOCs
