--- src/renderer/shared/hocs/withErrorToast.js
@@ -1,11 +1,0 @@
-import { connect } from 'react-redux';
-
-import { showErrorToast } from 'shared/actions/toastsActions';
-
-export default function withErrorToast(propName = 'showToast') {
-  const mapDispatchToProps = (dispatch) => ({
-    [propName]: (message) => dispatch(showErrorToast({ message }))
-  });
-
-  return connect(null, mapDispatchToProps);
-}
